pyproject.toml
[project]
name = "structural-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=0.1.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "markdown>=3.7",
    "openai>=1.58.1",
    "packaging>=24.2",
    "phidata>=2.7.5",
    "python-dotenv>=1.0.1",
    "python-markdown-math>=0.8",
]
